Fix issues with custom context menu items
